- Реализовал часть адаптера для разворачиваемых комментариев

+++ app/src/main/java/units/r00we/test_task/network/Issue.java
@@ -80,7 +80,7 @@ public class Issue {
         return user;
     }
 
-    public float getComments() {
+    public int getComments() {
         return comments;
     }
